+++ mods/sbz_base/init.lua
@@ -339,7 +339,7 @@ dofile(MP .. "/lag_delayer.lua")
 dofile(MP .. "/sbz_node_damage.lua")
 dofile(MP .. "/sbz_on_hover.lua")
 dofile(MP .. "/sbz_player_inside.lua")
-
+dofile(MP .. "/playtime_and_afk.lua")
 --vector.random_direction was added in 5.10-dev, but I use 5.9, so make sure this exists
 --code borrowed from builtin/vector.lua in 5.10-dev
 if not vector.random_direction then